==> 1.Toomanycells_result/Tcell/5.Cellcycle.ipynb <==
import scanpy as sc
import pandas as pd
import numpy as np

adata = sc.read_h5ad('./data/Tcell_PDL1.h5ad')

np.unique(adata.obs['Sub_Cluster'])

cell_cycle_genes = [x.strip() for x in open('./data/cell_cycle_vignette_files/regev_lab_cell_cycle_genes.txt')]
cell_cycle_genes

s_genes = cell_cycle_genes[:43]
g2m_genes = cell_cycle_genes[43:]
cell_cycle_genes = [x for x in cell_cycle_genes if x in adata.var_names]

sc.pp.normalize_total(adata, target_sum=1e6)

sc.pp.log1p(adata)
sc.pp.scale(adata)

sc.tl.score_genes_cell_cycle(adata, s_genes=s_genes, g2m_genes=g2m_genes)


adata.obs[['phase']].to_csv('./data/cellcycle_phase.csv',index=True,header=True)

for i in np.unique(adata.obs['Sub_Cluster']):
    ann = pd.read_csv('./result/cluster/'+i+'_cluster.csv',
                  index_col=0,header=0,sep=',',
                 usecols=['cell','cluster'])
    data = adata[adata.obs['Sub_Cluster']==i]
    data.obs = data.obs.join(ann)
    response = data.obs[['cluster','phase']]
    response.to_csv('./data/Cellcycle_count/'+i+'_cellcycle_count.csv',header=True,index=True)

adata.obs['item'] = adata.obs.index

adata.obs.index = adata.obs['item']

adata.obs.rename(columns={'phase':'label'},inplace=True)
np.unique(adata.obs['label'])

for i in np.unique(adata.obs['Sub_Cluster']) :
    response = adata[adata.obs['Sub_Cluster']==i].obs[['label']]
    response.to_csv('./data/'+i+'_cycle.csv',header=True, index=True, sep=',')
