==> 1.Toomanycells_result/Tcell/7.create_linechart_data.ipynb <==
import numpy as np
import pandas as pd
import scanpy as sc

sc.settings.verbosity = 3             # verbosity: errors (0), warnings (1), info (2), hints (3)
sc.logging.print_header()
sc.settings.set_figure_params(dpi=200, facecolor='white')

subcluster = 't_CD8-CXCL13'

adata = sc.read_h5ad('./data/Tcell_PDL1.h5ad')

adata = adata[adata.obs['Sub_Cluster']==subcluster]

adata

ann = pd.read_csv('./result/cluster/' + subcluster + '_cluster.csv',
                  index_col=0,header=0,sep=',',
                 usecols=['cell','cluster'])

adata.obs = adata.obs.join(ann,how='left')

adata.obs

df = adata.obs[['Patient', 'Response','Timeline']].value_counts().reset_index(name='count')[['Patient','Response','Timeline']]
df

for i in df.index:
    df.loc[i, 'count'] = len(adata[(adata.obs['Patient']==df.loc[i, 'Patient'])&(adata.obs['Timeline']==df.loc[i, 'Timeline'])])
df

adata.obs['cluster'] = adata.obs['cluster'].replace((12,13,16),0)

df['Res_Time'] = df.apply(lambda x: x['Response']+ '_'+ x['Timeline'],axis=1)

for i in df.index:
    df.loc[i, 'fraction'] = len(adata[(adata.obs['Patient']==df.loc[i,'Patient'])&(adata.obs['Timeline']==df.loc[i,'Timeline'])&(adata.obs['cluster']==0)])/df.loc[i,'count']
df

for i in df.index:
    df.loc[i,'fraction'] = df.loc[i, 'fraction'] * 100
df

df.to_csv('./data/lineplot/'+subcluster+'lineplot.csv',header=True,index=True)


df2 = df[['Response','Timeline']].value_counts().reset_index(name='count')[['Response','Timeline']]
df2

for i in df2.index:
    df2.loc[i, 'mean'] = np.mean(df[(df['Response']==df2.loc[i, 'Response'])&(df['Timeline']==df2.loc[i, 'Timeline'])]['fraction'])

df2.to_csv('/Users/liupeiwen/toomanycells/analysis_result/Tcell/ICB/'+subcluster+'_df/data_plot/mean_plot.csv',header=True,index=True)
