==> 1.Toomanycells_result/Tcell/9.spearman_exh_Sphase.ipynb <==
import scanpy as sc
import pandas as pd
import numpy as np

data = sc.read_h5ad('./data/Tcell_PDL1.h5ad')

np.unique(data.obs['Sub_Cluster'])

phase = pd.read_csv('./data/cellcycle_phase.csv',index_col=0)
phase

data.obs = data.obs.join(phase)
data

zhang = pd.read_table('./data/exhausted geneset.csv')
zhang

zhang = np.unique(zhang['Gene Symbol'])

sc.pp.normalize_total(data,target_sum=1e6)

for j in np.unique(data.obs['Sub_Cluster']):
    adata = data[data.obs['Sub_Cluster']==j]
    ann = pd.read_csv('./result/cluster/' + j + '_cluster.csv',
                  usecols=['cell','cluster'],index_col=0,header=0)
    adata.obs = adata.obs.join(ann)
    df = pd.DataFrame(data=np.unique(adata.obs['cluster']),columns=['Cluster'])
    
    for i in df.index:
        a=adata[adata.obs['cluster']==df.loc[i,'Cluster']]
            
        df.loc[i,'Count'] = len(a)
        df.loc[i,'G1_Count'] = len(a[a.obs['phase']=='G1'])
        df.loc[i,'S_Count'] = len(a[a.obs['phase']=='S'])
        df.loc[i,'G2M_Count'] = len(a[a.obs['phase']=='G2M'])

    
    df['G1_Prop'] = df['G1_Count'] / df['Count']
    df['S_Prop'] = df['S_Count'] / df['Count']
    df['G2M_Prop'] = df['G2M_Count'] / df['Count']
    
    adata[:,adata.var_names.isin(zhang)]
    
    for i in df.index:
        a = adata[adata.obs['cluster']==df.loc[i,'Cluster']]
        df.loc[i,'exhuast_score'] = sum(a[:,a.var_names.isin(zhang)].to_df().mean())    
        
    df.to_csv('./data/spearman_data/' + j + '_spearman_data.csv',index=False,header=True)

adata = data[data.obs['Sub_Cluster']=='t_CD8-CXCL13']
adata

ann = pd.read_csv('./result/cluster/t_CD8-CXCL13_cluster.csv',
                  usecols=['cell','cluster'],index_col=0,header=0)
ann

adata.obs = adata.obs.join(ann)

df = pd.DataFrame(data=np.unique(adata.obs['cluster']),columns=['Cluster'])

for i in df.index:
    a = adata[adata.obs['cluster']==df.loc[i,'Cluster']]
    df.loc[i,'Count'] = len(a)
    df.loc[i,'G1_Count'] = len(a[a.obs['phase']=='G1'])
    df.loc[i,'S_Count'] = len(a[a.obs['phase']=='S'])
    df.loc[i,'G2M_Count'] = len(a[a.obs['phase']=='G2M'])



df['G1_Prop'] = df['G1_Count'] / df['Count']
df['S_Prop'] = df['S_Count'] / df['Count']
df['G2M_Prop'] = df['G2M_Count'] / df['Count']

adata[:,adata.var_names.isin(zhang)]

a = adata[:,adata.var_names.isin(zhang)]

sum(a.to_df().mean())

sum(adata[:,adata.var_names.isin(zhang)].to_df().mean())

for i in df.index:
    a = adata[adata.obs['cluster']==df.loc[i,'Cluster']]
    df.loc[i,'exhuast_score'] = sum(a[:,a.var_names.isin(zhang)].to_df().mean())

df

df.to_csv('./data/spearman_data.csv',index=False,header=True)
